+++ integration/interactions/src/main/kotlin/com/splunk/rum/integration/interactions/InteractionsModuleIntegration.kt
@@ -36,7 +36,7 @@ import com.splunk.rum.integration.agent.internal.identification.ComposeElementId
 import com.splunk.rum.integration.agent.internal.module.ModuleIntegration
 import com.splunk.rum.integration.agent.internal.utils.runIfComposeUiExists
 import com.splunk.rum.integration.interactions.api.InteractionCapture
-import io.opentelemetry.android.instrumentation.InstallationContext
+import io.opentelemetry.api.OpenTelemetry
 import java.util.concurrent.TimeUnit
 
 internal object InteractionsModuleIntegration : ModuleIntegration<InteractionsModuleConfiguration>(
@@ -64,8 +64,8 @@ internal object InteractionsModuleIntegration : ModuleIntegration<InteractionsMo
     }
 
     override fun onInstall(
-        context: Context,
-        oTelInstallationContext: InstallationContext,
+        application: Application,
+        openTelemetry: OpenTelemetry,
         moduleConfigurations: List<ModuleConfiguration>
     ) {
         // No-op, as Interactions does not require any specific installation steps.