+++ integration/sessionreplay/src/main/kotlin/com/splunk/rum/integration/sessionreplay/SessionReplayModuleIntegration.kt
@@ -16,6 +16,7 @@
 
 package com.splunk.rum.integration.sessionreplay
 
+import android.app.Application
 import android.content.Context
 import android.webkit.WebView
 import com.splunk.android.instrumentation.recording.core.api.DataListener
@@ -35,7 +36,7 @@ import com.splunk.rum.integration.agent.internal.utils.runIfComposeUiExists
 import com.splunk.rum.integration.sessionreplay.api.SessionReplay as SplunkSessionReplay
 import com.splunk.rum.integration.sessionreplay.api.Status
 import com.splunk.rum.integration.sessionreplay.index.TimeIndex
-import io.opentelemetry.android.instrumentation.InstallationContext
+import io.opentelemetry.api.OpenTelemetry
 import io.opentelemetry.api.common.Attributes
 import io.opentelemetry.api.common.Value
 import java.util.concurrent.TimeUnit
@@ -61,8 +62,8 @@ internal object SessionReplayModuleIntegration : ModuleIntegration<SessionReplay
     }
 
     override fun onInstall(
-        context: Context,
-        oTelInstallationContext: InstallationContext,
+        application: Application,
+        openTelemetry: OpenTelemetry,
         moduleConfigurations: List<ModuleConfiguration>
     ) {
         Logger.d(TAG, "onInstall()")