+++ integration/slowrendering/src/main/kotlin/com/splunk/rum/integration/slowrendering/SlowRenderingModuleIntegration.kt
@@ -16,14 +16,14 @@
 
 package com.splunk.rum.integration.slowrendering
 
-import android.content.Context
+import android.app.Application
 import com.splunk.rum.common.logger.Logger
 import com.splunk.rum.instrumentation.slowrendering.internal.SlowRenderingInstrumentation
 import com.splunk.rum.integration.agent.common.module.ModuleConfiguration
 import com.splunk.rum.integration.agent.common.module.find
 import com.splunk.rum.integration.agent.internal.legacy.LegacySlowRenderingModuleConfiguration
 import com.splunk.rum.integration.agent.internal.module.ModuleIntegration
-import io.opentelemetry.android.instrumentation.InstallationContext
+import io.opentelemetry.api.OpenTelemetry
 
 internal object SlowRenderingModuleIntegration : ModuleIntegration<SlowRenderingModuleConfiguration>(
     defaultModuleConfiguration = SlowRenderingModuleConfiguration()
@@ -32,8 +32,8 @@ internal object SlowRenderingModuleIntegration : ModuleIntegration<SlowRendering
     private const val TAG = "SlowRendering"
 
     override fun onInstall(
-        context: Context,
-        oTelInstallationContext: InstallationContext,
+        application: Application,
+        openTelemetry: OpenTelemetry,
         moduleConfigurations: List<ModuleConfiguration>
     ) {
         Logger.d(TAG, "onInstall()")
@@ -49,7 +49,7 @@ internal object SlowRenderingModuleIntegration : ModuleIntegration<SlowRendering
             Logger.d(TAG, "Installing Slow Rendering Detector")
             val slowRenderingInstrumentation = SlowRenderingInstrumentation()
             slowRenderingInstrumentation.setSlowRenderingDetectionPollInterval(interval)
-            slowRenderingInstrumentation.install(context, oTelInstallationContext.openTelemetry)
+            slowRenderingInstrumentation.install(application, openTelemetry)
         } else {
             Logger.d(TAG, "Slow Rendering detection is disabled")
         }